Album and Song pages broked down into components
